requirements.txt
ansible==2.2.0.0
cffi==1.9.1
cryptography==1.7.1
Django==1.10.4
djangorestframework==3.5.3
enum34==1.1.6
idna==2.1
ipaddress==1.0.17
Jinja2==2.8
MarkupSafe==0.23
paramiko==2.1.1
pyasn1==0.1.9
pycparser==2.17
pycrypto==2.6.1
PyYAML==3.12
six==1.10.0
